requirements.txt
requests~=2.32.5
pytest~=9.0.2
python-dotenv~=1.2.1
allure-python-commons~=2.15.3